Fix space indexes
