+++ pkg/storage/fs/posix/decomposedfs/spaces_test.go
@@ -20,7 +20,6 @@ package decomposedfs_test
 
 import (
 	"context"
-	"os"
 
 	userv1beta1 "github.com/cs3org/go-cs3apis/cs3/identity/user/v1beta1"
 	cs3permissions "github.com/cs3org/go-cs3apis/cs3/permissions/v1beta1"
@@ -28,7 +27,6 @@ import (
 	provider "github.com/cs3org/go-cs3apis/cs3/storage/provider/v1beta1"
 	typesv1beta1 "github.com/cs3org/go-cs3apis/cs3/types/v1beta1"
 	ctxpkg "github.com/cs3org/reva/v2/pkg/ctx"
-	"github.com/cs3org/reva/v2/pkg/storage/fs/posix/decomposedfs"
 	"github.com/cs3org/reva/v2/pkg/storage/fs/posix/decomposedfs/node"
 	helpers "github.com/cs3org/reva/v2/pkg/storage/fs/posix/decomposedfs/testhelpers"
 	. "github.com/onsi/ginkgo/v2"
@@ -252,46 +250,6 @@ var _ = Describe("Spaces", func() {
 				})
 			})
 		})
-	})
-
-	Describe("ReadSpaceAndNodeFromSpaceTypeLink", func() {
-		var (
-			tmpdir string
-		)
-
-		BeforeEach(func() {
-			tmpdir, _ = os.MkdirTemp(os.TempDir(), "ReadSpaceAndNodeFromSpaceTypeLink-")
-		})
-
-		AfterEach(func() {
-			if tmpdir != "" {
-				os.RemoveAll(tmpdir)
-			}
-		})
-
-		DescribeTable("ReadSpaceAndNodeFromSpaceTypeLink",
-			func(link string, expectSpace string, expectedNode string, shouldErr bool) {
-				space, node, err := decomposedfs.ReadSpaceAndNodeFromIndexLink(link)
-				if shouldErr {
-					Expect(err).To(HaveOccurred())
-				} else {
-					Expect(err).ToNot(HaveOccurred())
-				}
-				Expect(space).To(Equal(expectSpace))
-				Expect(node).To(Equal(expectedNode))
-			},
-
-			Entry("invalid number of slashes", "../../../spaces/sp_ace-id/nodes/sh/or/tn/od/eid", "", "", true),
-			Entry("does not contain spaces", "../../../spac_s/sp/ace-id/nodes/sh/or/tn/od/eid", "", "", true),
-			Entry("does not contain nodes", "../../../spaces/sp/ace-id/nod_s/sh/or/tn/od/eid", "", "", true),
-			Entry("does not start with ..", "_./../../spaces/sp/ace-id/nodes/sh/or/tn/od/eid", "", "", true),
-			Entry("does not start with ../..", "../_./../spaces/sp/ace-id/nodes/sh/or/tn/od/eid", "", "", true),
-			Entry("does not start with ../../..", "../_./../spaces/sp/ace-id/nodes/sh/or/tn/od/eid", "", "", true),
-			Entry("invalid", "../../../spaces/space-id/nodes/sh/or/tn/od/eid", "", "", true),
-			Entry("uuid", "../../../spaces/4c/510ada-c86b-4815-8820-42cdf82c3d51/nodes/4c/51/0a/da/-c86b-4815-8820-42cdf82c3d51", "4c510ada-c86b-4815-8820-42cdf82c3d51", "4c510ada-c86b-4815-8820-42cdf82c3d51", false),
-			Entry("uuid", "../../../spaces/4c/510ada-c86b-4815-8820-42cdf82c3d51/nodes/4c/51/0a/da/-c86b-4815-8820-42cdf82c3d51.T.2022-02-24T12:35:18.196484592Z", "4c510ada-c86b-4815-8820-42cdf82c3d51", "4c510ada-c86b-4815-8820-42cdf82c3d51.T.2022-02-24T12:35:18.196484592Z", false),
-			Entry("short", "../../../spaces/sp/ace-id/nodes/sh/or/tn/od/eid", "space-id", "shortnodeid", false),
-		)
 	})
 
 	Describe("Update Space", func() {
